storage: factor out quiescer methods
Release note: None
